+++ Events/DisembarkEvent.cs
@@ -1,5 +1,6 @@
 ﻿using EddiDataDefinitions;
 using System;
+using System.Collections.Generic;
 using Utilities;
 
 namespace EddiEvents
@@ -56,7 +57,7 @@ namespace EddiEvents
         [PublicAPI("True if disembarking from your own ship")]
         public bool fromship => fromLocalId != null && !fromsrv && !fromtransport && !frommulticrew;
 
-        [PublicAPI("True if disembarking from an SRV")]
+        [PublicAPI( "True if disembarking from an SRV or Nomad" )]
         public bool fromsrv { get; } = fromSrv;
 
         [PublicAPI("True if disembarking from a transport ship (e.g. taxi or dropship)")]
@@ -75,5 +76,26 @@ namespace EddiEvents
         public int? fromLocalId { get; } = fromLocalId;
 
         public StationModel stationModel { get; } = stationModel;
+
+        public static bool Handle ( DateTime timestamp, string edType, string line, IDictionary<string, object> data, ref List<Event> events, bool fromLogLoad )
+        {
+            var fromSRV = JsonParsing.getBool(data, "SRV"); // true if getting out of SRV, false if getting out of a ship 
+            var fromTaxi = JsonParsing.getBool(data, "Taxi"); //  true when getting out of a transport ship (e.g. Apex Taxi or Frontline Solutions dropship)
+            var fromMultiCrew = JsonParsing.getBool(data, "Multicrew"); //  true when getting out of another player’s vessel
+            var fromLocalId = JsonParsing.getOptionalInt(data, "ID"); // player’s ship ID (if player's own vessel)
+
+            var system = JsonParsing.getString(data, "StarSystem");
+            var systemAddress = JsonParsing.getULong(data, "SystemAddress");
+            var body = JsonParsing.getString(data, "Body");
+            var bodyId = JsonParsing.getOptionalInt(data, "BodyID");
+            var onStation = JsonParsing.getOptionalBool(data, "OnStation");
+            var onPlanet = JsonParsing.getOptionalBool(data, "OnPlanet");
+
+            var marketId = JsonParsing.getOptionalLong(data, "MarketID");
+            EventParsing.StationNameAndType( data, out var stationName, out _, out var stationModel );  // if at a station
+
+            events.Add( new DisembarkEvent( timestamp, fromSRV, fromTaxi, fromMultiCrew, fromLocalId, system, systemAddress, body, bodyId, onStation, onPlanet, stationName, marketId, stationModel ) { raw = line, fromLoad = fromLogLoad } );
+            return true;
+        }
     }
 }