+++ Events/FighterRebuiltEvent.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using Utilities;
 
 namespace EddiEvents
@@ -15,5 +16,15 @@ namespace EddiEvents
 
         [PublicAPI("The fighter's id")]
         public int id { get; private set; } = id;
+
+        public static bool Handle ( DateTime timestamp, string edType, string line, IDictionary<string, object> data, ref List<Event> events, bool fromLogLoad )
+        {
+            if ( fromLogLoad ) { return true; } // Skip handling this during log loading
+
+            var loadout = JsonParsing.getString(data, "Loadout");
+            var fighterId = JsonParsing.getInt(data, "ID");
+            events.Add( new FighterRebuiltEvent( timestamp, loadout, fighterId ) { raw = line, fromLoad = fromLogLoad } );
+            return true;
+        }
     }
 }