+++ Events/TouchdownEvent.cs
@@ -23,37 +23,37 @@ namespace EddiEvents
         : Event( timestamp, NAME )
     {
         public const string NAME = "Touchdown";
-        public const string DESCRIPTION = "Triggered when your ship touches down on a planet's surface";
+        public const string DESCRIPTION = "Triggered when your ship or vessel touches down on a planet's surface";
         public const string SAMPLE = "{ \"timestamp\":\"2021-05-01T21:40:39Z\", \"event\":\"Touchdown\", \"PlayerControlled\":true, \"Taxi\":false, \"Multicrew\":false, \"StarSystem\":\"Nervi\", \"SystemAddress\":2518721481067, \"Body\":\"Nervi 2 a\", \"BodyID\":17, \"OnStation\":false, \"OnPlanet\":true, \"Latitude\":40.741577, \"Longitude\":65.081482 }";
 
-        [PublicAPI("The name of the star system where the ship has touched down")]
+        [PublicAPI("The name of the star system where the ship or vessel has touched down")]
         public string systemname { get; private set; } = system;
 
-        [PublicAPI( "The numeric system address of the star system where the ship has touched down" )]
+        [PublicAPI( "The numeric system address of the star system where the ship or vessel has touched down" )]
         public ulong systemAddress { get; private set; } = systemAddress;
 
-        [PublicAPI("The name of the body where the ship has touched down")]
+        [PublicAPI( "The name of the body where the ship or vessel has touched down" )]
         public string bodyname { get; private set; } = body;
 
-        [PublicAPI( "The numeric ID of the body where the ship has touched down" )]
+        [PublicAPI( "The numeric ID of the body where the ship or vessel has touched down" )]
         public int? bodyId { get; private set; } = bodyId;
 
-        [PublicAPI("The longitude from where the ship has touched down")]
+        [PublicAPI( "The longitude from where the ship or vessel has touched down" )]
         public decimal? longitude { get; private set; } = longitude;
 
-        [PublicAPI("The latitude from where the ship has touched down")]
+        [PublicAPI( "The latitude from where the ship or vessel has touched down" )]
         public decimal? latitude { get; private set; } = latitude;
 
-        [PublicAPI("True if the ship is a transport (e.g. taxi or dropship)")]
+        [PublicAPI( "True if the ship or vessel is a transport (e.g. taxi or dropship)" )]
         public bool? taxi { get; private set; } = taxi;
 
-        [PublicAPI("True if the ship is belongs to another player")]
+        [PublicAPI( "True if the ship or vessel is belongs to another player" )]
         public bool? multicrew { get; private set; } = multicrew;
 
-        [PublicAPI("True if the ship is controlled by the player")]
+        [PublicAPI( "True if the ship or vessel is controlled by the player" )]
         public bool playercontrolled { get; private set; } = playercontrolled;
 
-        [PublicAPI("The nearest location from where the ship has touched down")]
+        [PublicAPI( "The nearest location from where the ship or vessel has touched down" )]
         public string nearestdestination => nearestDestination.localizedName;
 
         [ PublicAPI( "True if the body is eligible for a first footfall" ) ]